requirements.txt
mysqlclient
djangorestframework
django-filter
djangorestframework-jwt